Update pre_migration_master.ps1
Fixed an issue where it would throw an error trying to stop the crawl monitoring process if it was never started.

+++ pre_migration_master.ps1
@@ -536,7 +536,9 @@ if ($mode -eq "single") {
 	if ($report -ne "") {
 		GeneratePostScanReport $source
 	}
-	Stop-Process $crawlMonitor
+	if ($notifications -eq "on") {
+		Stop-Process $crawlMonitor
+	}
 	sendNotification("Scan complete!")
 } elseif ($mode -eq "import") {
 	$crawlMonitor = $null
@@ -569,7 +571,9 @@ if ($mode -eq "single") {
 	if ($report -eq "overall") {
 		GeneratePostScanReport $rows
 	}
-	Stop-Process $crawlMonitor
+	if ($notifications -eq "on") {
+		Stop-Process $crawlMonitor
+	}
 	sendNotification("Scan complete!")
 
 } elseif ($mode -eq "report") {
